+++ gazebo_ros_control_plugin/CMakeLists.txt
@@ -1,6 +1,7 @@
-# http://ros.org/doc/groovy/api/catkin/html/user_guide/supposed.html
 cmake_minimum_required(VERSION 2.8.3)
-project(ros_control_gazebo_plugin)
+
+project(gazebo_ros_control_plugin)
+
 # Load catkin and all dependencies required for this package
 # TODO: remove all from COMPONENTS that are not catkin packages.
 find_package(catkin REQUIRED COMPONENTS roscpp gazebo_ros controller_manager
@@ -15,8 +16,8 @@ endif()
 link_directories(${GAZEBO_LIBRARY_DIRS})
 include_directories(include ${Boost_INCLUDE_DIR} ${catkin_INCLUDE_DIRS} ${GAZEBO_INCLUDE_DIRS})
 
-add_library(ros_control_gazebo_plugin src/ros_control_gazebo_plugin.cpp)
-target_link_libraries(ros_control_gazebo_plugin ${catkin_LIBRARIES})
+add_library(${PROJECT_NAME} src/ros_control_plugin.cpp)
+target_link_libraries(${PROJECT_NAME} ${catkin_LIBRARIES})
 
 # TODO: fill in what other packages will need to use this package
 ## DEPENDS: system dependencies of this project that dependent projects also need
